% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_concurrent_regression_band.R
\name{confint_concurrent}
\alias{confint_concurrent}
\title{Fast n Fair Simultaneous Confidence Band for the functional parameters of the concurrent regression model}
\usage{
confint_concurrent(
  concurrent_list,
  interval = "confidence",
  mse_scalar = "ub",
  time_vec = NULL,
  err_str = "t",
  nu0_hat = "singh_df",
  ...
)
}
\arguments{
\item{concurrent_list}{object returned from running concurrent regression with `fRegress_concurrent`}

\item{interval}{return a "confidence" or "prediction" band? "confidence" by default}

\item{mse_scalar}{should the variance estimator be unbiased ("ub") or maximum likelihood ("mle")?}

\item{time_vec}{sampling points of the functional observations; default (NULL) sets the sampling points as 1:T, where T is the number of rows in the y matrix returned in the concurrent list}

\item{err_str}{do the functional errors follow a Gaussian process ("normal") or Student's t process ("t")? "t" is default}

\item{nu0_hat}{either "singh_df" (default) or a numerical value specifying the degrees of freedom of the Student's t error process}

\item{...}{additional arguments that can be passed to `confidence_band`, which creates the final simultaneous band; of note, argument `conf.level` (default is 0.95) can be set to the desired confidence level for the simultaneous band}
}
\value{
object `simultaneous_band_list`, which is a list containing a simultaneous confidence/prediction band with `conf.int` confidence for each concurrent regression parameter
}
\description{
Fast n Fair Simultaneous Confidence Band for the functional parameters of the concurrent regression model
}
