% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffscbExtra-package.R
\docType{package}
\name{ffscbExtra-package}
\alias{ffscbExtra-package}
\title{ffscbExtra: Fast 'n' Fair Simultaneous Confidence and Prediction Bands for Functional Parameters including Concurrent Regression}
\description{
Simultaneous confidence and prediction bands for functional parameters including those in concurrent regression.
}
\author{
\strong{Maintainer}: Michael Creutzinger \email{creutzml@colostate.edu}

Authors:
\itemize{
  \item Dominik Liebl \email{r.packages.mail@gmail.com}
}

}
\keyword{internal}
