% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRegress_concurrent.R
\name{fRegress_concurrent}
\alias{fRegress_concurrent}
\title{Functional Concurrent Regression without any smoothing or regularization}
\usage{
fRegress_concurrent(y_mat, x_array, intercept = FALSE)
}
\arguments{
\item{y_mat}{matrix of response curves, with number of rows equal to the number of sampling points, T, and the number of columns equal to the number of observations, N (T x N)}

\item{x_array}{first two dimensions are the same as `y_mat`, while the third dimension represents the number of predictor variables, K (T x N x K)}

\item{intercept}{logical argument: does `x_array` contain a matrix of one's representing the intercept? `FALSE` by default, so the function creates the intercept term
\itemize{
     \item Note: if there's only one predictor variable aside from the intercept, `x_array` can be a T x N matrix with `intercept` set to `FALSE`
}}
}
\value{
a list containing the predicted functional observations, `y_hat_mat`, the estimated functional beta parameters, `beta_hat_mat`, the functional residuals, `y_resid_mat`, the original functional observations, `y_mat`, the original functional/scalar predictor variables, `x_array`, estimated functional DFFITS, `dffits_mat`, and estimated functional studentized residuals, `ext_stu_res_mat`
}
\description{
Functional Concurrent Regression without any smoothing or regularization
}
