% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalPrediction-package.R
\docType{package}
\name{FunctionalPrediction-package}
\alias{FunctionalPrediction}
\alias{FunctionalPrediction-package}
\title{FunctionalPrediction: Fast 'n' Fair Simultaneous Confidence and Prediction Bands for Functional Concurrent Regression}
\description{
Simultaneous confidence and prediction bands for functional concurrent regression.
}
\author{
\strong{Maintainer}: Michael L. Creutzinger \email{creutzml@colostate.edu}

Authors:
\itemize{
  \item Dominik Liebl \email{r.packages.mail@gmail.com}
  \item Julia L. Sharp
}

}
\keyword{internal}
