% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_concurrent_regression_band.R
\name{predict_concurrent}
\alias{predict_concurrent}
\title{Fast 'n' Fair Simultaneous Confidence/Preidction Band for Concurrent Functional Regression}
\usage{
predict_concurrent(
  concurrent_list,
  interval = "prediction",
  new_dat = NULL,
  time_vec = NULL,
  mse_scalar = "ub",
  err_str = "t",
  nu0_hat = "singh_df",
  ...
)
}
\arguments{
\item{concurrent_list}{object returned from running concurrent regression with `fRegress_concurrent`}

\item{interval}{return a "confidence" or "prediction" band?}

\item{new_dat}{structure should be a vector if the predictor variables are scalar, otherwise it should be a matrix if functional
\itemize{
     \item vector length should be Kx1, where K is the number of predictor variables, including the intercept.
     \item matrix should be TxK, where T is the number of sampling points and K is the number of predictor variables.
}}

\item{time_vec}{sampling points of the functional observations; default (NULL) sets the sampling points as 1:T, where T is the number of rows in the y matrix returned in the concurrent list}

\item{mse_scalar}{should the variance estimator be unbiased ("ub") or maximum likelihood ("mle")?}

\item{err_str}{do the functional errors follow a Gaussian process ("normal") or Student's t process ("t")? "t" is default}

\item{nu0_hat}{either "singh_df" (default) or a numerical value specifying the degrees of freedom of the Student's t error process}

\item{additional}{arguments that can be passed to `confidence_band`, which creates the final simultaneous band; of note, argument `conf.level` can be set to the desired confidence level for the simultaneous band}
}
\value{
a list containing fast n fair simultaneous confidence or prediction band for the marginal mean of a concurrent functional regression ("simultaneous_band") and an estimate of the degrees of freedom for the Student's t error process ("nu0_hat_est"), if applicable
}
\description{
Fast 'n' Fair Simultaneous Confidence/Preidction Band for Concurrent Functional Regression
}
